# src/main/frontend/public/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamstrack REST API
  version: 0.2.0
  description: |
    The HTTP API behind the Hamstrack web app — everything the UI does goes
    through it, and it is available to users for scripting and integrations.

    **Authentication** — call `POST /api/auth/login` and send the returned
    access token as `Authorization: Bearer <token>` (use the **Authorize**
    button here). Tokens expire after 24 h; refresh via the rotating
    `refresh_token` cookie with `POST /api/auth/refresh`.

    **Access model** — a resource you cannot see returns `404`, whether it
    does not exist or you are not a member of its workspace. Lists are
    currently unpaginated (beta). The API is unversioned while in beta;
    breaking changes are announced in release notes.

    **Rate limits** — sensitive auth endpoints share a per-IP budget
    (15 requests/min by default), and repeated failed logins for one account
    trigger an exponential backoff. Both return `429` with a `Retry-After`
    header (seconds).
  contact:
    name: Hamstrack
    url: https://github.com/Zherikhov/easyTask
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://hamstrack.com/api
    description: Cloud
  - url: /api
    description: This instance (Cloud or self-hosted DC)
...
security:
  - bearerAuth: []
paths:
...
  /admin/statuses:
    get:
      tags: [Admin]
      summary: List catalog statuses (admin)
      operationId: adminListStatuses
      responses:
        '200':
          description: Catalog with usage counts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      summary: Create a status (admin)
      operationId: adminCreateStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertStatus' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Name already exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
  /admin/statuses/{id}:
    patch:
      tags: [Admin]
      summary: Update a status (admin)
      operationId: adminUpdateStatus
      parameters: [ { $ref: '#/components/parameters/adminId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertStatus' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Delete a status, optionally remapping issues (admin)
      description: |
        409 while issues reference the status and `replaceWithId` is absent,
        or when the deletion would leave a workflow empty / the replacement
        is not part of every affected workflow. `POST /{id}/archive` and
        `/{id}/unarchive` (not listed individually) hide an entry from new
        use instead. The same delete/archive semantics apply to priorities
        and issue types.
      operationId: adminDeleteStatus
      parameters:
        - $ref: '#/components/parameters/adminId'
        - name: replaceWithId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: In use — remap required or blocked by integrity rules
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
  /admin/priorities:
    get:
      tags: [Admin]
      summary: List catalog priorities (admin)
      description: Full CRUD mirrors `/admin/statuses` — POST, PATCH `/{id}`, DELETE `/{id}?replaceWithId=`, POST `/{id}/archive|unarchive`.
      operationId: adminListPriorities
      responses:
        '200':
          description: Catalog with usage counts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminPriority' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/issue-types:
    get:
      tags: [Admin]
      summary: List catalog issue types (admin)
      description: Full CRUD mirrors `/admin/statuses`.
      operationId: adminListIssueTypes
      responses:
        '200':
          description: Catalog with usage counts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminIssueType' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/fields:
    get:
      tags: [Admin]
      summary: List catalog custom fields (admin)
      description: |
        POST creates a field (`key` blank = derived from the name; key and
        type are immutable afterwards — stored values depend on them), PATCH
        `/{id}` updates name/config/description, POST `/{id}/archive` and
        `/{id}/unarchive` hide a field from forms without touching values.
        DELETE `/{id}` refuses (409) while issues hold values for the field
        unless `?dropValues=true` is passed — the values are then dropped;
        there is no remap across arbitrary value shapes.
      operationId: adminListFields
      responses:
        '200':
          description: Catalog with usage counts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminField' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      summary: Create a custom field (admin)
      operationId: adminCreateField
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertField' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminField' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Name or key already exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422':
          description: Select field without options / option missing id or label
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
  /admin/field-sets:
    get:
      tags: [Admin]
      summary: List field sets (admin)
      description: |
        POST / PATCH `/{id}` take `{name, items:[{fieldId, required,
        showOnCreate}]}` (full replacement, display order; a required field
        is always shown on create); DELETE refuses while projects use the
        set. The system default set ("No fields") is empty and not deletable.
      operationId: adminListFieldSets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminFieldSet' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/workflows:
    get:
      tags: [Admin]
      summary: List workflows (admin)
      operationId: adminListWorkflows
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminWorkflow' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      summary: Create a workflow (admin)
      description: |
        `PATCH /admin/workflows/{id}` takes the same body (full replacement
        of statuses and transitions; 409 when a removed status still holds
        issues in bound projects). `DELETE /admin/workflows/{id}` refuses
        (409) while projects use the workflow; the system default is not
        deletable.
      operationId: adminCreateWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertWorkflow' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminWorkflow' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Name already exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
  /admin/statuses/{id}/usage:
    get:
      tags: [Admin]
      summary: Usage detail of a catalog entry (admin)
      description: |
        The expansion behind a usage chip — names instead of counts: the
        workflows/sets containing the entry and the projects that effectively
        use it through them (deduplicated; projects on a system default are
        included). The same endpoint exists for the other catalogs:
        `GET /admin/priorities/{id}/usage`, `GET /admin/issue-types/{id}/usage`,
        `GET /admin/fields/{id}/usage` (not listed individually).
      operationId: adminStatusUsage
      parameters: [ { $ref: '#/components/parameters/adminId' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/priority-sets:
    get:
      tags: [Admin]
      summary: List priority sets (admin)
      description: POST / PATCH `/{id}` take `{name, items:[{priorityId, isDefault}]}` (full replacement, display order); DELETE refuses while projects use the set.
      operationId: adminListPrioritySets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminPrioritySet' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/issue-type-sets:
    get:
      tags: [Admin]
      summary: List issue type sets (admin)
      description: |
        POST / PATCH `/{id}` take `{name, typeIds: […]}` (full replacement,
        display order; a set can never be empty); DELETE refuses while
        projects use the set, and the system default ("All types") is not
        deletable. A type set restricts only issue creation and type changes —
        existing issues keep a type that has left the set.
      operationId: adminListIssueTypeSets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminIssueTypeSet' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /api/auth/login` (24 h lifetime).
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating refresh token cookie set on login (30-day lifetime, path `/api/auth`).
  parameters:
...
    adminId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Forbidden:
      description: Authenticated but not allowed (insufficient role)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Not found — or not a member of the containing workspace
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
  schemas:
    ProblemDetail:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string, example: Conflict }
        status: { type: integer, example: 409 }
        detail: { type: string, example: Issue was modified by someone else — refresh and retry }
...
    Priority:
      type: object
      description: Catalog entity (was a closed enum before M1)
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: High }
        color: { type: string, example: '#EA580C' }
        icon: { type: string, nullable: true, example: chevron-up, description: lucide icon name }
        position: { type: integer }
    StatusCategory:
      type: string
      enum: [TODO, IN_PROGRESS, DONE]
      description: Drives board grouping and backlog filtering

    # ---- Admin console schemas ----
    UsageInfo:
      type: object
      description: Where a catalog entry is used; counts irrelevant to the type are 0
      properties:
        workflows: { type: integer, format: int64 }
        sets: { type: integer, format: int64 }
        projects: { type: integer, format: int64 }
        issues: { type: integer, format: int64 }
    UpsertStatus:
      type: object
      required: [name, category]
      properties:
        name: { type: string, maxLength: 100 }
        category: { $ref: '#/components/schemas/StatusCategory' }
        color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$', example: '#3B82F6' }
        position: { type: integer, nullable: true, description: Omitted = appended last }
    AdminStatus:
      allOf:
        - $ref: '#/components/schemas/StatusResponse'
        - type: object
          properties:
            archived: { type: boolean }
            usage: { $ref: '#/components/schemas/UsageInfo' }
    AdminPriority:
      allOf:
        - $ref: '#/components/schemas/Priority'
        - type: object
          properties:
            archived: { type: boolean }
            usage: { $ref: '#/components/schemas/UsageInfo' }
    AdminIssueType:
      allOf:
        - $ref: '#/components/schemas/IssueTypeResponse'
        - type: object
          properties:
            archived: { type: boolean }
            usage: { $ref: '#/components/schemas/UsageInfo' }
    UpsertWorkflow:
      type: object
      required: [name, statusIds]
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, nullable: true }
        statusIds:
          type: array
          description: Board-column order
          items: { type: string, format: uuid }
        transitions:
          type: array
          items: { $ref: '#/components/schemas/TransitionRule' }
    AdminWorkflow:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Dev workflow }
        description: { type: string, nullable: true }
        systemDefault: { type: boolean, description: Used by projects without a binding; not deletable }
        statuses:
          type: array
          items: { $ref: '#/components/schemas/StatusResponse' }
        transitions:
          type: array
          items: { $ref: '#/components/schemas/TransitionRule' }
        projectsUsing: { type: integer, format: int64 }
    AdminPrioritySet:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Default priorities }
        systemDefault: { type: boolean }
        items:
          type: array
          items:
            type: object
            properties:
              priority: { $ref: '#/components/schemas/Priority' }
              isDefault: { type: boolean, description: Default for new issues }
        projectsUsing: { type: integer, format: int64 }
...
    AdminIssueTypeSet:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: All types }
        systemDefault: { type: boolean }
        types:
          type: array
          description: Display order
          items: { $ref: '#/components/schemas/IssueTypeResponse' }
        projectsUsing: { type: integer, format: int64 }
    UsageDetail:
      type: object
      description: '"Where exactly is this used?" — names instead of counts'
      properties:
        workflows:
          type: array
          items: { type: string }
        sets:
          type: array
          items: { type: string }
        projects:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              key: { type: string, example: DEMO }
              name: { type: string }
        issues: { type: integer, format: int64 }
    FieldType:
      type: string
      enum: [TEXT, TEXTAREA, NUMBER, DATE, SELECT, MULTI_SELECT, USER, CHECKBOX, URL]
      description: >-
        Determines the JSON value shape: TEXT/TEXTAREA/URL string ·
        NUMBER number · DATE "YYYY-MM-DD" string · SELECT option id ·
        MULTI_SELECT array of option ids · USER user UUID (workspace
        member) · CHECKBOX boolean
    FieldConfig:
      type: object
      nullable: true
      description: '`options` for selects; `min`/`max` for numbers'
      properties:
        options:
          type: array
          items:
            type: object
            required: [id, label]
            properties:
              id: { type: string, example: critical, description: Referenced by stored values — stable once used }
              label: { type: string, example: Critical }
              color: { type: string, nullable: true, example: '#B91C1C' }
        min: { type: number, nullable: true }
        max: { type: number, nullable: true }
    UpsertField:
      type: object
      required: [name, type]
      properties:
        name: { type: string, maxLength: 100 }
        key:
          type: string
          pattern: '[a-z0-9_]*'
          maxLength: 50
          description: Blank on create = derived from the name; immutable afterwards
        type: { $ref: '#/components/schemas/FieldType' }
        config: { $ref: '#/components/schemas/FieldConfig' }
        description: { type: string, maxLength: 500, nullable: true, description: Shown as a hint in issue forms }
    AdminField:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string, example: story_points }
        name: { type: string, example: Story points }
        type: { $ref: '#/components/schemas/FieldType' }
        config: { $ref: '#/components/schemas/FieldConfig' }
        description: { type: string, nullable: true }
        archived: { type: boolean }
        usage: { $ref: '#/components/schemas/UsageInfo' }
    AdminFieldSet:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Engineering fields }
        systemDefault: { type: boolean }
        items:
          type: array
          description: Display order
          items:
            type: object
            properties:
              field: { $ref: '#/components/schemas/AdminField' }
              required: { type: boolean, description: Must be filled on create; cannot be cleared }
              showOnCreate: { type: boolean, description: Rendered in the create form }
        projectsUsing: { type: integer, format: int64 }
...
    IssueTypeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Bug }
        color: { type: string, example: '#EF4444' }
        icon: { type: string, nullable: true, example: bug }
        position: { type: integer }
    StatusResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: In Progress }
        color: { type: string, example: '#3B82F6' }
        category: { $ref: '#/components/schemas/StatusCategory' }
        position: { type: integer }
    TransitionRule:
      type: object
      description: fromStatusId null = "from any status"
      properties:
        fromStatusId: { type: string, format: uuid, nullable: true }
        toStatusId: { type: string, format: uuid }
...